Pipeline refactor: Introduce domain package
* Refactor constant to workflow domain module
* Separate string constant to config module
* Remove un-use FastQReadType
