+++ data_processors/pipeline/lambdas/tests/test_orchestrator.py
@@ -7,7 +7,7 @@ from mockito import when
 
 from data_portal.models import Workflow
 from data_portal.tests.factories import WorkflowFactory, TestConstant, SequenceRunFactory
-from data_processors.pipeline.constant import WorkflowType, WorkflowStatus
+from data_processors.pipeline.domain.workflow import WorkflowType, WorkflowStatus
 from data_processors.pipeline.lambdas import orchestrator
 from data_processors.pipeline.tests import _rand
 from data_processors.pipeline.tests.case import logger, PipelineUnitTestCase, PipelineIntegrationTestCase
