+++ data_processors/pipeline/lambdas/workflow_update.py
@@ -16,7 +16,8 @@ import logging
 from data_portal.models import Workflow
 from data_processors.pipeline.services import workflow_srv, notification_srv
 from data_processors.pipeline.lambdas import wes_handler
-from data_processors.pipeline.constant import WorkflowType, SQS_NOTIFICATION_QUEUE_ARN
+from data_processors.pipeline.domain.config import SQS_NOTIFICATION_QUEUE_ARN
+from data_processors.pipeline.domain.workflow import WorkflowType
 from utils import libjson, libssm, libsqs
 
 logger = logging.getLogger()
